Added retry functionality
Added the ability for developers to say that a question can be redone if
the user got the answer wrong, or they simply wish to try the question
again anyway.  This is a per question option, therefore the developer
can decide to have only select questions retryable, or all questions.